account-automation/requirements.txt
azure-identity
msgraph-core
msgraph-sdk
requests
python-dotenv